+++ flecsi/run/control.hh
@@ -231,20 +231,18 @@ private:
     Output a graph of the control model.
   */
 #if defined(FLECSI_ENABLE_GRAPHVIZ)
-  int write() const {
+  void write(const std::string & p) const {
     flecsi::util::graphviz gv;
     point_writer::write(registry_, gv);
-    std::string file = program() + "-control-model.dot";
+    std::string file = p + "-control-model.dot";
     gv.write(file);
-    return flecsi::run::status::control_model;
   } // write
 
-  int write_sorted() const {
+  void write_sorted(const std::string & p) const {
     flecsi::util::graphviz gv;
     point_writer::write_sorted(sort(), gv);
-    std::string file = program() + "-control-model-sorted.dot";
+    std::string file = p + "-control-model-sorted.dot";
     gv.write(file);
-    return flecsi::run::status::control_model_sorted;
   } // write_sorted
 #endif
 
@@ -333,17 +331,19 @@ public:
     Execute the control model. This method does a topological sort of the
     actions under each of the control points to determine a non-unique, but
     valid ordering, and executes the actions.  If the policy `P` inherits from
-    `control_base`, an object of type \c P is value-initialized and destroyed
-    before this function returns.
+    `control_base`, an object of type \c P is initialized from \a aa and
+    destroyed before this function returns.
     \c control_base::exception can be thrown for early
-    termination. \return code from a thrown \c control_base::exception or the
+    termination.
+    \param aa only if inheriting from \c control_base
+    \return code from a thrown \c control_base::exception or the
     bitwise or of return values of executed actions.
    */
-
-  static int execute() {
+  template<class... AA>
+  static int invoke(AA &&... aa) {
     if constexpr(is_control_base_policy) {
       try {
-        P pol = P();
+        P pol(std::forward<AA>(aa)...);
         return instance().run(&pol);
       }
       catch(control_base::exception e) {
@@ -351,23 +351,34 @@ public:
       }
     }
     else {
+      static_assert(!sizeof...(aa), "arguments allowed only with control_base");
       return instance().run(nullptr);
     }
-  } // execute
+  }
+  /// Call \c #invoke with no arguments.
+  /// \deprecated Use \c invoke directly or call \c runtime::main.
+  [[deprecated("use invoke")]] static int execute() {
+    return invoke();
+  }
 
   /*!
     Process control model command-line options.
-    \param s initialization status
+    \param s initialization status from \c initialize
     \return status of control model output if requested, else \a s
+    \deprecated Use \c runtime::main.
+    \see arguments::action::operation
    */
-
-  static int check_status(int s) {
+  [[deprecated("use flecsi::runtime")]] static int check_status(int s) {
 #if defined(FLECSI_ENABLE_GRAPHVIZ)
+    // If a confused client calls this without having called initialize,
+    // argv0 will be empty, which is a mild form of failure.
     switch(s) {
       case flecsi::run::status::control_model:
-        return instance().write();
+        write_graph(argv0);
+        break;
       case flecsi::run::status::control_model_sorted:
-        return instance().write_sorted();
+        write_actions(argv0);
+        break;
       default:
         break;
     } // switch
@@ -375,7 +386,40 @@ public:
     return s;
   } // check_status
 
+#ifdef FLECSI_ENABLE_GRAPHVIZ
+  static void write_graph(const std::string & p) {
+    instance().write(p);
+  }
+  static void write_actions(const std::string & p) {
+    instance().write_sorted(p);
+  }
+#endif
 }; // struct control
+
+struct call_policy : control_base {
+  enum control_points_enum { single };
+  using control_points = list<point<single>>;
+  struct node_policy {};
+
+  template<class F>
+  explicit call_policy(F && f) : f(std::forward<F>(f)) {}
+  int operator()() const {
+    return f();
+  }
+
+private:
+  std::function<int()> f;
+};
+
+inline const char *
+operator*(call_policy::control_points_enum) {
+  return "single";
+}
+
+/// A trivial control model that calls a single function.
+/// Its control policy object can be constructed from any callable with the
+/// signature `int()`.
+using call = control<call_policy>;
 
 /// \}
 } // namespace run