+++ tutorial/.auxiliary/control-overview/control-overview.cc
@@ -126,18 +126,9 @@ const auto dep_bn = action_b.add(action_n);
 
 int
 main(int argc, char ** argv) {
-  auto status = flecsi::initialize(argc, argv);
-  status = control::check_status(status);
-
-  if(status != flecsi::run::status::success) {
-    return status < flecsi::run::status::clean ? 0 : status;
-  }
-
-  flecsi::log::add_output_stream("clog", std::clog, true);
-
-  status = flecsi::start(control::execute);
-
-  flecsi::finalize();
-
-  return status;
+  flecsi::run::arguments args(argc, argv);
+  const flecsi::run::dependencies_guard dg(args.dep);
+  const flecsi::runtime run(args.cfg);
+  flecsi::flog::add_output_stream("clog", std::clog, true);
+  return run.main<control>(args.act);
 } // main